Aggregating CpG probes to genes

Function aggregate_probes aggregates CpG probes to corresponding genes.

aggregate_probes(data, keep = NULL,
  genom.data = illumina_humanmethylation_27_data, genes.col = 11,
  probes.col = 1)

Arguments

data
data frame containing methylation values for CpG probes. Columns correspond to probes, rows to samples.

keep
the name of the column or vector of columns names we want to keep.

genom.data
data frame which contains information about CpG probes and corresponding genes, by default in our package we use illumina_humanmethylation_27_data.

genes.col
number of column in genom.data containing informations about genes (genes symbols).

probes.col
number of column in genom.data containing informations about probes (probes symbols).

Value

A data frame with CpG probes aggregated to genes. If there were more than one probe corresponding to a gene, value is a mean of those probes.

Examples

# NOT RUN {
library(MLExpRessodata)
BRCA_genes <- aggregate_probes(BRCA_methylation_all, keep="SUBTYPE")
# }