Results from test for two datasets.

Function calculate_comparison_table produces a dataset containing p-values and folds from tests evaluated on two datasets e.g methylation or expression. In addition, it produces an importance ranking column, which is the geometric mean of p-values from both tests and a column with a number of probes related to the gene.

calculate_comparison_table(data1, data2, condition1, condition2, test1, test2,
  genom.data = MLExpResso::illumina_humanmethylation_27_data,
  genes.col = 11)

Arguments

data1
First dataset, suitable for test1.

data2
Second dataset, suitable for test2.

condition1
Condition for first dataset.

condition2
Condition for second dataset.

test1
Type of test for first dataset.

test2
Type of test for second dataset.

genom.data
Data frame which contains information about CpG probes and corresponding genes, by default in our package we use illumina_humanmethylation_27_data.

genes.col
Number of column in genom.data containing informations about genes (genes symbols).

Value

Data frame containing logatithm of fold and p-values from chosen tests.

See also

calculate_test

Examples

# NOT RUN {
library(MLExpRessoData)
condition_exp <- ifelse(BRCA_exp$SUBTYPE == "LumA","LumA", "other")
condition_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA", "other")

BRCA_methylation_gen <- aggregate_probes(BRCA_met)

data_met <- BRCA_methylation_gen
data_exp <- BRCA_exp
compare <- calculate_comparison_table(data_exp,data_met,cond_exp,cond_met, "nbinom2", "ttest")
# }