Basic statistics for gene.

Function gene_stats generate a data frame containing basic statistics for chosen gene.

gene_stats(data, condition, gene, round = 2)

Arguments

data
data frame containing genes expression or methylation mapped to gene.

condition
vector of levels of groups coresponding to order of samples in data.

gene
symbol of chosen gene.

round
num

Value

data frame with following columns:

min

minimal value for chosen gene in group

1st Q

first quartile

med

median of gene values

mean

mean for gene values

3rd Q

third quartile

max

maximal value for chosen gene in group