Illumina_humanmethylation_27_data

The variables are as follows:

Name. CG number from CG database (format cg########).

GenomeBuild. Genome build.

Chr. Chromosome on which the target locus is located.

MapInfo. Cancer subtype.

Source. Genomic position of C in CG dinucleotide.

SourceVersion. Genomic position source.

SourceSeq. Original sequence of the region covered by assay probes.

TSS_Coordinat. Transcription start site genomic coordinate.

Gene_Strand. Gene strand.

Gene_ID. RefSeq identifier (GeneID).

Symbol. Gene Symbol.

Synonym. Gene synonyms.

Accession. Gene Accession (this is the accession of the longest transcript).

GID. GI ID.

Annotation. Gene annotation from NCBI database.

Product. Gene product description from NCBI database.

Distance_to_TSS. Distance of CG dinucleotide to transcription start site.

CPG_ISLAND. oolean variable denoting whether the loci is located in a CpG island (by relaxed definition) .

CPG_ISLAND_LOCATIONS. Chromosomal location and genomic coordonates of the CpG island from NCBI database.

MIR_CPG_ISLAND. Chromosome:start-end of upstream CPG island from a micro RNA.

MIR_NAMES. Name of micro RNA near locus.

Format

data.frame