Methylation path on chosen gene.

Function plot_methylation_path visualise a chosen gene with marked CpG probes. It shows the mean methylation level for each probe in group. Also we can exact the line corresonding to gene. In this case we see what are the locations of probes on gene in HG18 coordinates. We can as well draw a locations of CpG islands.

plot_methylation_path(data, condition, gene, show_gene = FALSE,
  observ = FALSE, islands = TRUE, title = TRUE, ...)

Arguments

data
data frame containing values from methylation: columns corespond to CpG probes, rows to samples.

condition
vector of levels corresponding to order of samples in data.

gene
name of chosen gene.

show_gene
logical. If TRUE line corresopnding to gene will be draw.

observ
logical. If TRUE dots corresponding to CpG probes will be draw.

islands
logical. If TRUE line corresopnding to islands should be draw.

title
logical. If TRUE title saying what gene we visualise will be add.

...
other parameters.

Value

Object of class ggplot containing visualisation of methylation on gene.

See also

plot_gene

Examples

# NOT RUN {
library(MLExpRessodata)
condition_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA", "other")
plot_methylation_path(BRCA_met, condition_met, "BRCA2")
# }