t test for methylation per each gene

Function test_tstudent computes statistics and p-values for methylation data mapped to genes.

test_tstudent(data, condition, ...)

Arguments

data
data frame containing values of methylation: columns coresponds to genes, rows to samples.

condition
vector of levels coresponding to order of samples in data.

Value

A data frame with the following columns:

id

The ID of the observable, taken from the row names of the counts slots.

mean

The base mean.

log.fold

The log of the fold change.

pval

The p-values for rejecting the null hypothesis about the means equality.

padj

The adjusted p-values.

t.stat

The value of the t-test statistic.

B

The log-odds that that gene is differentially expressed.

See also

expr_nbinom